Fix a whole bunch of things
